+++ components/content/ContentAlert.vue
@@ -8,7 +8,7 @@ defineProps({
 </script>
 
 <template>
-  <div class="p-4 rounded-md bg-yellow-50 dark:bg-lime-700">
+  <div class="alert-box p-4 rounded-md bg-yellow-50 dark:bg-lime-900">
     <div class="flex">
       <div v-if="icon" class="flex-shrink-0 mr-3">
         <SiteIcon
@@ -21,14 +21,20 @@ defineProps({
         <!-- eslint-disable-next-line vuejs-accessibility/heading-has-content -->
         <h3
           v-if="$slots.title"
-          class="m-0 text-base font-medium text-yellow-800 dark:text-yellow-100"
+          class="alert-box__title m-0 text-base font-medium text-yellow-800 dark:text-yellow-100"
         >
           <ContentSlot :use="$slots.title" unwrap="p" />
         </h3>
-        <div class="mt-2 text-sm text-yellow-700 dark:text-yellow-100">
+        <div class="alert-box__content mt-2 text-sm text-yellow-700 dark:text-yellow-100">
           <ContentSlot :use="$slots.default" unwrap="p" />
         </div>
       </div>
     </div>
   </div>
 </template>
+
+<style lang="postcss">
+:is(.dark body .prose .alert-box :not(h1, h2, h3, h4, h5, h6) a) {
+  @apply text-lime-100;
+}
+</style>