+++ nuxt.config.ts
@@ -28,7 +28,7 @@ export default defineNuxtConfig({
     },
     highlight: {
       // See the available themes on https://github.com/shikijs/shiki/blob/main/docs/themes.md#all-theme
-      theme: "dracula",
+      theme: "github-dark",
       preload: ["py", "css", "scss", "ts", "json", "js", "jsx"],
     },
   },